survivor-predictor/requirements.txt
anthropic>=0.40.0
praw>=7.7.0
youtube-transcript-api>=0.6.0
streamlit>=1.30.0
rich>=13.0.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.18.0
tweepy>=4.14.0
atproto>=0.0.40
python-dotenv>=1.0.0
pyyaml>=6.0.0
